refactor: extract daemon bootstrap client
