+++ Cargo.toml
@@ -2,6 +2,7 @@
 members = [
     "crates/atm-core",
     "crates/atm",
+    "crates/atm-daemon-client",
     "crates/atm-graft",
     "crates/sc-lint-attributes",
     "crates/sc-lint-boundary",
