+++ boundaries/atm-daemon-client/daemon-bootstrap.toml
@@ -1,0 +1,66 @@
+boundary_id = "BOUNDARY-DaemonBootstrapClient"
+owner_package = "atm-daemon-client"
+owner_crate_path = "atm_daemon_client"
+name = "DaemonBootstrapClient"
+
+[public]
+facade = "SameHostDaemonBootstrap"
+
+[implementation]
+type = "DaemonSupervisor"
+module = "atm_daemon_client"
+visibility = "public"
+constructor = "public"
+
+[composition]
+roots = [
+  "atm::composition",
+  "atm_graft",
+]
+
+[ownership]
+io_owns = [
+  "same_host_daemon_bootstrap_types",
+  "daemon_launch_gate",
+  "daemon_auto_start_spawn",
+]
+io_forbidden = [
+  "sqlite",
+  "daemon_request_dispatch",
+  "graft_session_runtime",
+]
+
+[dependencies]
+allowed_dependents = ["atm", "atm-graft"]
+allowed_dependencies = ["atm-core"]
+forbidden_edges = [
+  "atm-daemon-client -> atm-daemon",
+  "atm-daemon-client -> atm-rusqlite",
+]
+
+[references]
+scope = "outside_owner_crate"
+forbidden = ["atm_daemon::client", "rusqlite::Connection"]
+
+[contracts]
+request_types = []
+response_types = []
+error_types = ["AtmError"]
+
+[testing]
+allowed_test_double_paths = []
+forbidden_test_bypasses = ["atm_daemon::client"]
+
+[enforcement]
+lint_rules = ["LINT-BOUNDARY-DAEMON-BOOTSTRAP-DEPENDENCIES"]
+review_gates = [
+  "no_bootstrap_to_daemon_edge",
+  "no_bootstrap_to_sqlite_edge",
+]
+
+[status]
+state = "active"
+notes = [
+  "shared same-host daemon bootstrap helpers extracted from atm and atm-graft",
+  "transport adapters remain in their owning crates; this crate owns only launch/bootstrap helpers",
+]
