+++ crates/atm-daemon-client/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "atm-daemon-client"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+description = "Shared same-host ATM daemon bootstrap helpers for thin clients."
+repository.workspace = true
+homepage.workspace = true
+
+[dependencies]
+atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "1.2.0" }
+fs2 = "0.4"
