+++ crates/atm-graft/Cargo.toml
@@ -11,7 +11,7 @@ homepage.workspace = true
 
 [dependencies]
 atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "1.2.0" }
-fs2 = "0.4"
+atm-daemon-client = { path = "../atm-daemon-client", version = "1.2.0" }
 interprocess = "2.4.2"
 tracing.workspace = true
