+++ crates/atm/Cargo.toml
@@ -20,6 +20,7 @@ path = "src/main.rs"
 
 [dependencies]
 atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "1.2.0" }
+atm-daemon-client = { path = "../atm-daemon-client", version = "1.2.0" }
 anyhow.workspace = true
 chrono = { version = "0.4", features = ["serde"] }
 clap = { version = "4", features = ["derive"] }
@@ -31,7 +32,6 @@ serde_json.workspace = true
 time = "0.3"
 tracing.workspace = true
 tracing-subscriber.workspace = true
-fs2 = "0.4"
 
 [dev-dependencies]
 atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "1.2.0", features = ["test-utils"] }
